Don't check out a commit if it is already checked out
When building a stack of commits, the merger would end up checking
out each commit again in the series, even if it had just created
that commit and therefore that commit was already checked out.

Detect that case and leave the working tree as-is to speed up
building refs for large stacks of changes.

Change-Id: Ib9347c0c4f887b7d00020a61fa831c637de8562e