web: Allow clients (browsers, proxies) to cache status.json
Follows-up aa4f2e7. While we're sending Last-Modified already,
that's purely informational and doesn't do much.

In theory proxies could figure it out by observing or polling
the backend (with rate limiting) but afaik that isn't generally
implemented.

Web browsers typically send If-Modified-Since. So for requests
coming from the users directly, Response.conditional_response_app
handles the If-Modified-Since header and returns early with
HTTPNotModified if needed (and no body content).

In addition Cache-Control/Expires headers allow clients/proxies
to handle it themselves (without even a 304 rountrip). This is
preferred and informs the client of the actual expiry instead of
having it ask us every time whether or not a timestamp is too old.

Response:
 200 OK
 Cache-Control: public, must-revalidate, max-age=(seconds)
 Expires: (RFC 2822 timestamp + seconds)
 Last-Modified: (RFC 2822 timestamp)

Request:
 GET
 If-Modified-Since: (RFC 2822 timestamp)

Response:
 304 Not Modified
 Expires: (RFC 2822 timestamp + seconds)

http://docs.webob.org/en/1.1/modules/webob.html

Change-Id: I51f31f9d7965d805e147fda4070feead528601ac