Status: Don't raise "abort" as error to the user
If the request was aborted, the failure should not be presented
to the user as nothing actually went wrong.

This happens when a request is not finished when next 'update'
starts. This is amplified by the page visibility events that
will start an 'update' loop prematurely.

Without this, switching between tabs often causes the user to
see a "status.json: abort" message when they come back to the
dashboard.

Also remove redundant (and incorrectly labeled) arguments.

Change-Id: I834384535f8d04d1274336bc993c20035be01d60