Update patch set 2
Patch Set 2:

It's not just the write() that acquires a lock (which was moved to a try/catch in this patch). It turns out that the config_writer() method also tries to acquire this lock.

With this patch applies, we still hit a premature build failure:
https://phabricator.wikimedia.org/T86730#1215139

Again, this happens whenever Zuul cancels Jenkins builds (or when Jenkins jobs are aborted for others reasons) and happens to interrupt zuul-cloner. Even with the workspace gitclean/reset turned on, internal git directories are left alone.

Patch-set: 2