Update patch set 1
Patch Set 1: Code-Review-1

I really don't like using regex to substitute the correct username.

If we want mergers to use different accounts to the zuul server then the merger client should not pass across the URL to clone but rather the project details so that the merger server can clone based off what it knows to be gerrit (url, creds etc).

I think I'd support such a change as it would neaten up the merger server a little. However, I'm not sure it's something we need or want to support.

The real fix here would be to update the documentation to state that the merger must use the same credentials as the server.

Is there a reason for using different ones?

Patch-set: 1
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1