Update patch set 1
Patch Set 1: Code-Review-1

The existing behavior of zuul-cloner appears to be correct. indicated_branch/zuulref then indicated branch are always prefered over master.

This is because we only attempt to use master in the case where the indicated branch does not exist. So as long as the branch exists it will be used.

Feel free to propose an update to the comment though as it could more clearly state that master is only considered when indicated branch does not exist.

Patch-set: 1
Label: Code-Review=-1