Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Unless I'm missing something you appear to be changing the actual cloner strategy. Your earlier commit documented this was the case in the commit message, but now you say you're only updating the comment? Isn't the strategy still changing?

If the strategy is changing - to what is probably more correct - we clearly haven't tested this bit of the cloning very well. Would we be able to see some tests that prove the correct order/strategy is done?

Patch-set: 6