Update patch set 2
Patch Set 2: Code-Review-1

This is actually supported in 2.4 and possibly earlier.  The gerrit system in question seemed to have an error and was not providing it for some reason.  In the case that gerrit is not functioning properly, I don't think that Zuul should hide that.  If we wanted to check for this and log an error I think that would be fine, but we should not proceed in any case.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1