Update patch set 3
Patch Set 3: Code-Review-1

Thanks.  This looks good, but I think we should have a test for this.  Possibly one where jobs are configured to run in a pipeline, but filtering causes no actual jobs to match for a given change.  Mostly I would like to verify that in this case, the change is ultimately removed from the pipeline even though it does not report.

I think it's fine if the test is added in a followup change, but I would like to wait on merging this until it exists.

Also, I could see a potential use case for someone wanting reports with no jobs (perhays a gating system where there are no jobs that run for certain kinds of changes).  We may want a configuration option to handle that case.  But we don't have it right now, so we can defer that for later.

Patch-set: 3