Update patch set 1
Patch Set 1:

Yeah I'll definitely write tests for this eventually, but to be honest I don't think the current implementation of Zuul reporting in this series of patches is anywhere ready for this patch.

The problem is that most of the actually formatting for reporting takes place in the scheduler and most of that assumes a gerrit or smtp reporting approach. The scheduler should probably only pass pure data structures to the reporters and allow the reporters (gerrit, smtp, github, etc) to format those data structures according to the type of message being sent.

This would allow the github reporter (for example) to use github-flavored markdown rather than relying on HTML jumbled together by the scheduler.

Patch-set: 1
Reviewer: Gerrit User 13487 <13487@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0