Update patch set 21
Patch Set 21: Code-Review-1

I worked on a github reporter implementation and found the current reporter interface lacking. As described in my comment on I3c34bb1354adb7c360e173c227f00bf987b7d30e, I believe the "Reporter" objects should not only be responsible for relaying messages to their connection objects, but also for actually formatting the contents of the message according to the medium they are reporting to.

In this implementation there is still a great deal of "reporting" behavior that still happens in the BasePipelineManager class found in the scheduler module, in particular this is where formatting happens non-optionally. In my view, formatting should be left to the reporter. Of course this can be delegated to a shared formatting class for reporters who use the same message format as seems to be the case with gerrit and smtp.

Patch-set: 21