Add the ability to abort jobs on a patch merge
For slow running 3rd party CI systems, a patch can merge
before the CI system had a chance to run it. It will run eventually,
but its result is not very useful. Add an option to abort these
running jobs and remove them from the pipeline so that resources
can be allocated to changes that are not yet merged.

Change-Id: I64531ddd926ffdc91a8b018f6184470922192372