cloner: make more zuul args optional
When running in a timer job ZUUL_BRANCH and ZUUL_REF are not
provided.  Rather than having to set those in the job definition,
just make them optional for zuul-cloner.  The prepare logic is
updated to avoid incorrect string substitutions in case they are
not supplied.

With those two optional, the only remaining required argument is
ZUUL_URL.  This is only required if ZUUL_REF is set.

Also, enhance debugability by always printing the commit sha, even
if cloner has simply checked out a branch.

Change-Id: I8f88cb8d1e49255d38c2c75dd226b3027779aa9e