Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I disagree with the premise, but I'm fine with this as a feature.

Specifically I believe the job results to still be useful even after the patch has merged. CI operators should be watching which runs fail and why (and compared to upstream). If there was a legitimate bug the CI system would have caught given time it's still useful to know that so that a new fix to correct the regression can be pushed.

Patch-set: 3
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1