Update patch set 2
Patch Set 2: Code-Review-1

Why not put a git checkout of every project that you depend on in the local mirror?  Even if Zuul doesn't list it in ZUUL_CHANGES, the actual build/pypi upload is delayed, so what you are testing will vary (if foo is in ZUUL_CHANGES because of change 42 then a git checkout will be used, but in a subsequent run after change 42 is merged but before its build is uploaded to pypi, the local mirror will not be used, and foo will be at least one revision behind).  Pypi uploads can be significantly delayed (8 hours is not unheard of) because they are a lower priority than tests.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1