Update patch set 2
Patch Set 2:

I agree that both should be available.  However, this doesn't let you implement either of them, this implements a third option, which is to conditionally test some unreleased versions of libraries when you would normally test released ones depending on which changes happen to be in flight at that moment.  I'm not sure that's something that we should support -- I don't think i could adequately explain to someone why a job which is supposed to test released versions of libraries is, in some conditions, not actually doing so.

Patch-set: 2