Update patch set 1
Patch Set 1:

By mass changing our labels in layout.yaml we had a bug that approvals are no more matched.

BaseFilter.matchesRequiredApprovals() does the matches based on the approval description normalized to replace spaces with dashes and in lower case:

 >>> import zuul.model
 >>> zuul.model.normalizeCategory('Code-Review')
 'code-review'
 >>>

Thus with an approval Code-Review: 2, the filter code reject the approval:

 if (normalizeCategory(approval['description']) != k or
    int(approval['value']) not in v):
    found_approval = False

Because 'code-review' != 'Code-Review'.

Maybe the matches should be done against "type" and with both versions normalized.

Reference:
https://phabricator.wikimedia.org/T106436

Patch-set: 1