Dependencies from undefined projects
3rd party CI layout usually has only a few projects defined,
so it's possible that some changes depend on projects
which are unknown to Zuul scheduler.
These items had None as a "item.change.project", which
is not handled in many places, for ex. in reconfiguration.

These cases could be handled by defining these projects in layout
as "dummies": no jobs, no other non-standard attributes.
but this is obviously not a long-term solution.

We propose to use "default constructor", i.e. Project(name)
instead of None in Zuul scheduler,
so that such undefined changes will be handled more properly.

Change-Id: I7912197fb86c1a7becb7f43ca36078101f632715