Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

This helps a lot, thanks.  I do still think it's important that we define the behavior, and if we want to have dynamic project creation, that we actually add those projects to the layout so that they are singletons, and make sure that behavior is consistent everywhere.

If we decide to implement that, we may need to put getProject in the context of a pipeline so that we can associate it with a source, that way the source can verify if it exists.

Patch-set: 2
Label: Code-Review=-1