Jobs can now be configured to re-run if they fail
Enables jobs to be told to automatically re-run themselves
a configurable number of times, if they are known to fail for
temporary reasons unrelated to the code that they are being
used to test. No retries are enabled by default, so this does
not require any configuration changes to retain the status quo.

Change-Id: I503270967b8393901e93e7a0c74e179a3eff7fc0