Update patch set 3
Patch Set 3:

(1 comment)

Thanks for the clarification, that part looks good to me now.

Another question, however.. If a change references an unknown project to zuul for a dependent pipeline (eg gate) won't this change get things stuck? Perhaps for dependant projects unknown to zuul we need to enqueue a noop job. Although that runs the risk of zuul submitting back to gerrit for a project it shouldn't control. This may need some more thought..

Patch-set: 3