Update patch set 2
Patch Set 2: Code-Review-1

I am actually going to -1 this having given it some more thought. I don't think this is a feature that we want in zuul for aforementioned reasons. I'm only -1, not -2 so I'm happy to be out-voted.

If a job is truly unstable it shouldn't be in a voting pipeline. Put it in a non-reporting or experimental pipeline and re-trigger it as appropriate. Perhaps using the zuultrigger even.

Patch-set: 2
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1