Update patch set 3
Patch Set 3:

(1 comment)

Ah okay. I actually wonder if that's a bit of a bug. Where two projects don't have a common job and therefore don't automatically have a common change-queue but still need to be tested together because project A pulls in project B which it has a Depends-On for. I wouldn't expect this change to fix that though.


Also an inline comment about the commit message.

Patch-set: 3