Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

Can we see a test for this scenario in both a check type pipeline and a gate type pipeline?

eg make sure no jobs are launched - or if there were jobs that they were cancelled still

I'm also still a little unclear on what happens if there is a cross-project dependency in a dependant pipeline. Will it ignore the dependency? Good to see a test for this too.

Patch-set: 5
Label: Code-Review=-1