Update patch set 10
Patch Set 10:

So the layouts in tests/fixtures/layouts/ are prefixed with good_ or bad_. There is a test (test_layoutvalidator.py) that loads the layouts and validates them with the expected output to be the prefix. If we could add a layout to that folder that'd be great.

Patch-set: 10