Move Item formatting into Reporters
This will allow a reporter to decide how to handle the results of
each item. It can use the common plain text formatter (as has been
the case, '_formatItemReport') or it may generate a report itself.

This will be useful for the MySQL reporter where it will want to
create an entry in a table for each build.

Change-Id: Ib270334ff694fdff69a3028db8d6d7fed1f05176