Optionally require a zuul-ref when cloning
In some cases we may want to fail cloning instead of falling back
to a branch. One case is when cloning in gate tests, if we do not
fetch our zuul ref we should fail instead of running tests on the
development branch (which has already passed the gate tests).

Change-Id: I410c75b452684cc25e232990e9adeb7d77ccc0c7