Improve job dependencies using graph instead of tree
The Zuul configuration layout.yaml cannot handle dependency graphs.
Consider the following configuration:
gate:
  - A:
    - B:
      - D
      - E
    - C:
      - E

Zuul currently reports two E jobs in the status.json, but only one will be
started. This is because BuildSet in model.py index jobs by their names
and getBuild(job_name) is used to find jobs to run, as well as in many
other places in the code. This means that the job E will run once, when
either B or C has finished successfully.

Change the logic to register each job only once, which will avoid the
problem with duplicated entries in the status.json. Also change the logic
from OR to AND. This allows for specifying any generic dependency graph,
i.e. all acyclic graphs.

A simple pattern to specify a full dependency graph is then:
gate:
  - A:
    - B
    - C
  - B:
    - D
    - E
  - C:
    - E

In this graph, the job E will not start until both B and C have finished.

If possible, a check in the layout validator to avoid circular
dependencies would be nice, but the validator does not seem to resolve
regular expressions on the job names as done in Layout.getJob(name).

Fixes bug #1166937.

Signed-off-by: Fredrik Medley <fredrik.medley@autoliv.com>
Signed-off-by: Fredrik Medley <fredrik.medley@gmail.com>
Change-Id: I921940cafeea0738c39deb99357cfd7c91592359