Update to statsd 3.0
With statsd 3.0, the way to get a client object that is populated from
environment variables has changed from being a module import, i.e.

 import statsd

to a variable import

 from statsd.defaults.env import statsd

However, the test-cases as writen need to reload "statsd" (the module)
so that the tests dynamically pick up the host/port for the fake
statsd server started by the base test class (i.e. overriding any
values passed externally, such as from tox).

Instead of dealing with this transition via messy module reloads and
munging of global state, this change moves an explicit instantiation
of a statsd client into each object that will produce a stat.  We add
a small helper to set-up "self.stats" within these objects from the
environment.

Thus whatever STATSD_[HOST|PORT] is set to when you instantiate the
object will be the location stats are sent to.

The main reason for this change isto make the statsd requirement the
same as nodepool (done in I4791c9d26f2309f78a556de42af5b9945005aa46)
so the two can co-exist better when in distro packages.

Change-Id: Ib84655378bdb7c7c3c66bf6187b462b3be2f908d