Use 127.0.0.1 rather than localhost for statsd host
When we use "localhost" as the statsd server, statsd can possibly
choose ::1 (i.e. ipv6) and all stats tests fail because the fake stats
server is ipv4.

This seems more prevalent with python2.6 (failures in our centos6 CI,
replicated with Debian stable), probably because ordering of results
changes in later glibcs.

Anyway, specify 127.0.0.1 so we're getting the right thing

Change-Id: I2f821834ab4d7e06cdb5933b79b9f650d9668c0e