Update patch set 1
Patch Set 1:

Thanks James and Khai!

The syntax is very similar to GraphViz but the renderer is written in Python.  The upstream website is worth looking at: http://blockdiag.com/ . It provides a wide range of diagram types including sequence or networks layout.  I love it.

Regarding cycles, the doc is rather thin.  If another change enhances it with some examples, I will be happy to review/add the associated blockdiag that illustrate it.  That might be as simple as:

 A -> B -> C -> A

Patch-set: 1
Label: Verified=0