Update patch set 2
Patch Set 2:

(1 comment)

Many times on dowstream we face questions of what are the types of metrics, specially the per-second and accumulated keys are.

We thought that providing this introductory documentation would help with that, specially showing the calculated keys for timers and where to find each.

Would it make sense to have this introduction to types as well as a link to upstream documentation?

Patch-set: 2
Reviewer: Gerrit User 18158 <18158@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0