Update patch set 38
Patch Set 38:

(1 comment)

I fixed this on the plane and missed your patch.

I've done it slightly differently. The main problem is that if multiple gerrits are connected we need to know which sshkey to use for which. As such, if you use the same zuul.conf on your mergers you can look it up off the connection name. Therefore the merge job will pass along the connection name as a parameter. sshkey is used for legacy still.

Patch-set: 38
Label: Verified=0