Update patch set 39
Patch Set 39:

My view was that the ssh key is being used in the gerrit connection for a different purpose than in the merger. Sort of a "role" based access - to distinguish between scheduler and merger.

For instance for github, I don't need a ssh key at all in the scheduler. But I need it for the merger to clone private repositories.

If you have different gerrit instances, you could still set different access rights for the specific key(s).So, something like the the following:
 [zuul]
 sshkey = /var/lib/zuul/ssh/shceduler_key

 [merger]
 sshkey = /var/lib/zuul/ssh/merger_key

Then you can set up on the gerrit instances:
 Gerrit A -> scheduler_key: allow to run gerrit stream-events
                            allow to run gerrit review
             merger_key: allow to clone all repositories
 Gerrit B -> scheduler_key: allow to run gerrit stream-events
                            forbid to run gerrit review
             merger_key: allow to clone only some repositories
                         (not sure if this makes sense, but illustrates well)

Patch-set: 39
Label: Verified=0