Update patch set 39
Patch Set 39:

We could have a merger_sshkey and scheduler_sshkey assigned to the same gerrit connection.

I think for the moment assuming the merger has the same config as the scheduler is the easiest way forward. Another change further down the stack could allow mergers to use different credentials, but I'm not sure we need to add it here.

Patch-set: 39
Label: Verified=0