Log errors on zuul validating schema
When an invalid zuul layout.yaml is sent, zuul fails silently,
starting scheduler and not reconfiguring properly, and skipping
start of RPC and Webclient.
The only way to get aware of this problem is to execute the server
from command line to properly visualize the error.
Enclose this validation under an exception, and log the triggered
exception, so we can be aware of the failures with zuul layout.yaml

Change-Id: Iaae42935988a72ebe1534edf6c63d5cc2326f7b7