Update patch set 42
Patch Set 42:

So I think I've fixed the unit test you added. Basically after an event is triggered, if it can't be found in the source it is ignored. I had to tweak your test a little to represent this so you might want to double check I've got the edge case that you highlighted.

Patch-set: 42
Label: Verified=0