Update patch set 45
Patch Set 45:

Move the creation of the ssh wrapper back to zuul.merger.Merger. It belongs there and it fixes a bug which happened because the zuul.merger.Server was trying to create the ssh wrapper before instantiating zuul.merger.Merger which should create the parent directory of the wrappers.

Also remove the explicit creation of the parent directory from the test and leave it up to the merger. This would've exposed the bug earlier.

Patch-set: 45
Reviewer: Gerrit User 15215 <15215@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0