Update patch set 7
Patch Set 7: Code-Review-1

(3 comments)

Cool stuff, thanks!

The inline comments are just nit-picks if you'd like to address them.

My -1 is because I'd like to see one test for the flat and deep graphs to make sure they are equivalent. For example, have a generic

  def _test_dependeny_graph_dispatch_jobs_once(self, type):..

then have

  def test_dependeny_graph_dispatch_jobs_once_deep(self):
    _test_dependeny_graph_dispatch_jobs_once('deep')

etc.

See the CRD tests that compare functionality in check to gate.

This is to validate the flat/deep work the same. In fact, I wonder if a test to compare the two graphs would be feasible?

Patch-set: 7
Label: Code-Review=-1