Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Actually there is a lot here that will affect the connection work. That's fine if this is important and needs to go in first though.

-1 for inline comment.

I'd also like to get James' opinion on this.

I'm not a huge fan of duplicating the wrapper code either.

Patch-set: 2
Label: Code-Review=-1