Handle a float in assertReportedStat
Upstream have "future-proofed" [1] timing() results by making them a 6
digit float [1]; this breaks string matching when updating statsd
(Ib84655378bdb7c7c3c66bf6187b462b3be2f908d).

This modifies assertReportedStat to work in all cases by
special-casing timing results and doing a float comparision.

It also seemed a bit odd that "value" has to include a "kind" in the
string, but the argument "kind" is there when you want to ignore
"value".

So document the function and modify the value to just be a value and
use the "kind" passed in, and modify callers.

[1] https://github.com/jsocol/pystatsd/commit/f5ddc98d7ce20e02ba70a0d4543612f4fd209643

Change-Id: I4ba9fca1ba9e8904881ad1402ab2143abdc56896