Update patch set 24
Patch Set 24: Code-Review-1

(2 comments)

Some things inline, but also some general observations:

This is getting sophisticated enough that just using alembic from the start might be a good idea.  It's not too hard; see gertty for a simple, direct example.

We have mysql and postgres servers available on our test nodes, we should probably make our unit test use them directly.  See nodepool for an example.  We should not use sqlite _at all_ unless we actually expect it to be used in production because it is so different from the others.

On that note, we should seriously consider whether we want to support more than one database.  If no one is using this with postgres, we should not claim to support it.  If we do want to support it, we should understand that it is twice the work (development, testing, and support).  SQLAlchemy is a great ORM that happens to support multiple backend databases.  It is not a great database abstraction layer (in anything but the most trivial cases, such a thing does not really exist).  Perhaps this is trivial enough, I just don't want us to over-promise.

Patch-set: 24
Label: Code-Review=-1