Update patch set 26
Patch Set 26: Code-Review-1

Ah, I didn't realize you wanted to be able to have different pipelines report to different build tables (possibly even under different circumstances).  That merits more thought.

I don't understand why there are two initial alembic migrations.

We should really talk more about sqlite.  I'm worried we are about to make the same mistake every other project has made in trying to support more than one database with radically different semantics.  Note that you can still run unit tests locally if you have mysql installed.  This is what we do with nodepool.

I am definitely not suggesting that we abandon sqlalchemy -- we definitely should us it.  My argument is that sqlalchemy is *not* a database abstraction layer, and that it should not be used in an attempt to support more than one database from the same code base.  Rather, it is an ORM, and we should use it because it is a good ORM.  It *can* also be used to support multiple databases, but at the cost of multiple times the amount of work.  Every supported database needs to be tested, and have its own schema and migrations.  My argument is that we will only ever use at most, one -- MySQL -- so we should only test and support that one.

Anyway, let's chat about this with Clint and Monty who are fresh from doing similar work.

Patch-set: 26
Label: Code-Review=-1