Connection names for legacy configs
It has been set like that:
connections = {'_legacy_gerrit': Connection(name='gerrit')}

This leads to that in merger.py it creates ".ssh_wrapper_gerrit" file,
basing it on key from connections dict,
but tries to use ".ssh_wrapper__legacy_gerrit", basing on connection name.

I propose to set it to "_legacy_gerrit" everywhere (and the same for smtp).

Change-Id: I2c2d8bee7bcf98c5f12809f77b34e206d07b8438