Update patch set 1
Patch Set 1:

The harm is that I don't know what they are supposed to do, but this gets called every time a reconfiguration happens.  I'm pretty sure that whatever someone might choose in the future to put in here is *not* something that we want to have happen every time zuul reconfigures itself, which is going to be *a lot* more often in v3.

In general, I would really like to avoid adding stub code that's not actually used.  This just feels like we added tech debt.

Patch-set: 1