Add extra hardening on success-pattern url
Report can have parameters from a job substituted into the message.
If a reporter fails for whatever reason it is logged to zuul and
otherwise ignored quietly to the user.

If bad varaibles are attempted to be substituted into a message
log a warning and fallback to the result url for the reporter to
continue.

Change-Id: I521cfbb5873973014c43f9780722d2f80a7c12f0