Update build params before getting swift instruct
The swift instructions use the LOG_PATH build param in generating
the destination for logs. A custom build params function may have
updated how the LOG_PATH looks, so the swift instructions should be
done after the custom function in case LOG_PATH has changed.

This has the disadvantage that a custom function can't update any
of the swift instruction BUILD PARAMS. This shouldn't matter as
all of the swift instruction defaults are configurable on a per-job
basis.

Change-Id: I07f45c1c6cd4f61a61d3746b891a50dd51744f50