Update patch set 29
Patch Set 29:

> > Might we consider a streaming service similar to what Gerrit
 > provides?
 > 
 > I personally am ok with that solution. Do you have in mind
 > implementing that as an alternative or as an only option?

Well I think we *could* implement both options.  The webhook option would probably be easier for people using github webhooks today to integrate with zuul, so there's that.  I'd be curious to hear opinions form infra-admin's which they'd prefer as the implementation of choice.  To achieve streaming we need to setup an intermediary service that polls checking for new events and then broadcasting them out to any connected ssh client.  That makes the infra slightly more complex, but keeps it more consistent without we do it today with Gerrit.

Patch-set: 29
Label: Verified=0