Add ansible as a dependency
Note that ansible is licensed under the GPL which has implications
for the distribution of Zuul as a whole.

The dependency is on the 2.0 branch which should be released *very*
soon at which point it can change to a normal pypi dep.

Actually use ansible to launch a hello galaxy test.

Change-Id: I30f7ed2cadbb7d2381ec477e229d19c9c6e41743