Explicitly set basepython=python2.7
Without a basepython declaration (explicitly or implicitly in "py27",
etc venvs), tox uses the current version of python with which "tox"
itself was run.  On Debian (at least), this is now python3.  Many of the
venvs in this tox.ini require python2.7, however.

This change adds explicit basepython=python2.7 declarations to all the
venvs that require it.  Note the "docs" venv _does_ work with
python3 (at least with my quick local test with python3.4).

Change-Id: I8fd6827abed93ccfe2967f9495e63bdcd83ca71d