Set logserver variables using temp url key only
When Send-Temp-Url-Key is set to False and X-Account-Meta-Temp-Url-Key
is defined, there is no need for a swiftclient connection and the
zuul_swift_upload.py works using the provided temp_url_key.

Without this, the [swift] configuration section always requires OS_
settings, even if temp_url_key and logserver_prefix are already defined.

Change-Id: I8990359d47e9d5065d34af3faea2d24e63e736b8