Update patch set 2
Patch Set 2:

I've gotten as far as determining that the change queue for the "needed change" (change A) is None in the check that fails because its project is not the same Project object as change B.  I think that may mean that something is amiss with the project cache -- I think it may have moved from the scheduler to a connection which gets recreated on reconfiguration.  If that's the case, I think project objects (and the cache of them) should live through reconfiguration events.

Patch-set: 2
Label: Verified=0