Update patch set 4
Patch Set 4:

recheck

I'm not sure if only using a change id was done by design or not. Once a change has merged the changeid is mostly meaningless in the git history. However a UUID can be searched across multiple repos with some quick scripting. In other words, I'm not sure if this is a feature we want.

Patch-set: 4
Label: Verified=0