Update patch set 1
Patch Set 1:

(5 comments)

Hi Joshua, I appreciate your feedback and thanks for taking the time to review.

One thing I noticed with Zuul documentation though is many of the details seem to have implied documentation rather than explicit documentation. As someone who recently figured out how to setup Zuul and had to fight through it for a few days I feel it would be nice if many of the sections had more description of the purpose of the sections rather than assuming the user could figure it out and hope we can add more details to each section in the future.

For example I had no idea what gearman was and whether I needed both the gearman & gearman_server sections or not. I think omitting details in the [zuul] section is not helpful as this is making the assumption that a list of requirements is all a user needs to get started with Zuul. I feel that each section should at least have a small snippet (ideally a detailed one) describing the purpose of the section so that someone new to zuul has a fighting chance of understanding what they are configuring and why. As developers we often forget that the user reading the documentation does not always share the same knowledge as us.

I hope this proposal is helpful as I'd like to contribute more to this documentation assuming my submissions are accepted as I think this is useful tool that needs more documentation.

Patch-set: 1
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0