Add timeout/retries to zuul cloner
This adds timeout and retry arguments to the zuul-cloner command.
If a git clone takes longer than <timeout> seconds, it is aborted.
Git clone operations are retried <retries> times if they encounter
an error.

This only applies to cloning for the moment, but could potentially
be extended to fetches, etc.

Change-Id: Idf2a072426319079a7683fd74bb8bced13b5669f