Update patch set 12
Patch Set 12:

Rebased. To ease the transition for Openstack infrastructure I have made the argparse argument --cache-no-hardlinks to default to the environment variable ZUUL_CACHE_NO_HARDLINKS

This way, you can have Zuul to inject the parameter ZUUL_CACHE_NO_HARDLINKS=1 and that would have the same effect as passing --cache-no-hardlinks. Thus you can set the parameter now, and when this patch start being deployed it is back compatible.

I think it is safer than trying to sync this patch deployment and having to update all the Jenkins jobs to pass --cache-no-hardlinks.

Patch-set: 12