Update patch set 10
Patch Set 10:

Joshua, I think that is my test that was racy somehow. It killed the Gearman server and maybe that has various side effects on other tests.

Did another recheck, we might want a few more and if all pass that seems to me the code change itself is not racy and that the test suite I wrote was.

Patch-set: 10