Move gerrit logic from source to connection
Sources and connections are very tightly coupled in reality.  Rather
than trying to maintain them as two abstractions, consider the
connection to hold all of the information and logic about the reality
of the external resource it represents.  Make sources mere local
data structures that are used to interface a connection with a pipeline.

As seen in subsequent changes, this will allow us to simplify the
interconnections between objects.

Change-Id: I2dd88e1165267e4f987a205ba55923eaec7ea9ce