Remove scheduler parameter from connection registry
The connection registry should not have to know about the scheduler,
rather, the inverse is true.

(NB, connections themselves still know about the scheduler, but
that's okay, that happens after the connection registry is created.)

Drivers should be able to access the global configuration when
being created, so store that when the connection registry configures
itself.

Change-Id: Iea4b8fe3888b5eefd3df9ce385225b885f2caa0b