Update patch set 5
Patch Set 5:

I don't think this is what is needed to work in post/release pipelines.  It's quite easy for us to set ZUUL_BRANCH=$ZUUL_REF in project-config as a workaround. Checking out a branch doesn't ensure we end up in the same state as gerrit-git-clone on post jobs.

I've been pursuing a fix to this as well recently, and came to a different conclusion of what needs to be supported to function on pipelines that are triggered by ref-updated events.  My initial patch is here and looking for feedback before I push tests on it and finalise it.
https://review.openstack.org/#/c/293194/

Patch-set: 5
Reviewer: Gerrit User 12612 <12612@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0