Update patch set 3
Patch Set 3:

> It's starting to look like we don't actually care about the charset
 > here and are simply happy to defer to webob.  That makes this seem
 > like an overzealous test, and we should just remove the assertion.

I think a large portion of these asserts are overzealous, but that will require looking into what WebOb automatically adds to the headers and what zuul code is manually manipulating.  I'm happy to look into this if it is seen as necessary, but I'm also not sure it provides value moving forward.

I do however think the assert on 'content-type' being 'application/json' is a valid and important test as it is manually set in zuul code, and getting that right is important.  The charset just got removed to make WebOb more correct going forward.

I did not go for assertIn('application/json') as it's important that 'content-type' starts with 'application/json' exactly. I could loosen it up by accepting anything after a ';' though, as far as I can tell, there aren't any parameters that are valid to follow application/json, so we are just testing that for backwards compatibility.

Patch-set: 3