Update patch set 14
Patch Set 14:

The change from patch set 12 to 14 is that jobs are reported to status.json in the same order as before, i.e. the order the jobs are mentioned in the layout configuration. In patch set 12 the jobs were just sorted by name. (The only issue reported so far by my company.)

All tests still passes but as "testrepository.subunit was > 50 MB of uncompressed data", Jenkins reports failure.

Patch-set: 14
Label: Verified=0