Fix projects entering pipeline without jobs
Changes for a project which doesn't define builds for a specific
pipeline can still enter this pipeline. In this case the start message
gets posted to gerrit and the zuul-merger is triggered. After that
zuul notices that there are no builds to process.

The test for this is from Joshua Hesketh in
https://review.openstack.org/#/c/302800/

Change-Id: Ifa5d3b4f46adf90d39f2b5a551c1bf2bbb3137d8