Projects without jobs still enter pipeline
A project without any jobs for a particular pipeline may still enter
it if the trigger is matchd. For example, a project may be added
to 'check', have the start message posted to gerrit, have the
mergers do their thing and then it is decided that that there are
no jobs so it bails.

This is just a test to prove the bug. We shouldn't consider a
project for a pipeline if it hasn't got any jobs.

Change-Id: I9d8a0c5d14010fad022c46e63f37474fe1e18621