Projects without jobs still enter pipeline
Previously a project without any jobs for a particular pipeline may
still enter it if the trigger is matched. For example, a project may
be added to 'check', have the start message posted to gerrit, have the
mergers do their thing and then it is decided that that there are no
jobs so it bails.

This behavour is fixed in Ifa5d3b4f46adf90d39f2b5a551c1bf2bbb3137d8
and this change adds the tests.

Change-Id: I9d8a0c5d14010fad022c46e63f37474fe1e18621