Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure if we should bother with this. With the work in zuulv3 if this causes a merge conflict it's not worth the effort of resolving it. Instead we should fix the typo in the v3 branch.

Happy to be out voted though.

Patch-set: 1
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1